backend/requirements.txt
adrf==0.1.7
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aisuite==0.1.6
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.8.1
async-property==0.2.2
attrs==24.2.0
bcrypt==4.2.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1
django-apscheduler==0.6.2
django-cors-headers==4.4.0
django-filter==24.3
django-oauth-toolkit==2.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.10.0
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.5
idna==3.8
inflection==0.5.1
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
marshmallow==3.22.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.58.1
orjson==3.10.7
packaging==24.1
pandas==2.2.2
paramiko==3.5.0
pip==24.2
ply==3.11
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rpds-py==0.20.0
setuptools==72.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sqlparse==0.5.1
tenacity==8.5.0
tiktoken==0.7.0
tokenizers==0.21.0
tqdm==4.66.5
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
wheel==0.44.0
yarl==1.9.4
